--- tests/Unit/Services/LanguageIdentifierEnglishTest.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace Tests\Unit\Services;
-
-use Termorize\Services\LanguageIdentifier;
-use Tests\TestCase;
-
-class LanguageIdentifierEnglishTest extends TestCase
-{
-    /**
-     * @test
-     */
-    public function test()
-    {
-
-        $text = 'hello';
-        $textLang = LanguageIdentifier::identify($text);
-
-        $this->assertEquals('en', $textLang);
-
-    }
-}
